Separate useKeyboard from Form component
