+++ frontend/app/component/form/data_entry/recounted/RecountedForm.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import { useRecounted } from "@kiesraad/api";
 import { t } from "@kiesraad/i18n";
-import { BottomBar, Button, ChoiceList, Feedback, Form, KeyboardKey, KeyboardKeys } from "@kiesraad/ui";
+import { BottomBar, Button, ChoiceList, Feedback, Form, KeyboardKey, KeyboardKeys, useKeyboard } from "@kiesraad/ui";
 
 interface FormElements extends HTMLFormControlsCollection {
   yes: HTMLInputElement;
@@ -15,6 +15,8 @@ interface RecountedFormElement extends HTMLFormElement {
 
 export function RecountedForm() {
   const formRef = React.useRef<RecountedFormElement>(null);
+
+  useKeyboard(formRef);
 
   const getValues = React.useCallback(() => {
     const form = formRef.current;
