+++ frontend/lib/ui/Form/index.ts
@@ -1,2 +1,3 @@
 export * from "./FormLayout";
 export * from "./Form";
+export * from "./useKeyboard";
